cephadm: test_networks: Move tests from test_cephadm
